move the conversion function to a separate module
