+++ docs/scripts/generate_latitude_diff_table.py
@@ -1,10 +1,15 @@
 import pathlib
+import site
+
+site.addsitedir(str(pathlib.Path(__file__).parent))
+
 import re
 import sys
 
 import numpy as np
 import pandas as pd
 import tabulate
+from conversions import convert
 
 a = 6378137
 f_ = 298.257223563
@@ -12,27 +17,6 @@ f = 1 / f_
 e = np.sqrt(2 * f - f**2)
 
 b = a * (1 - f)
-
-
-def q(ϕ, e):
-    # from USGS Professional Paper 1395, "Map projections – A working manual" by John P. Snyder
-    return (1 - e**2) * (
-        np.sin(ϕ) / (1 - e**2 * np.sin(ϕ) ** 2)
-        - (1 / (2 * e)) * np.log((1 - e * np.sin(ϕ)) / (1 + e * np.sin(ϕ)))
-    )
-
-
-def authalic_radius(a, e):
-    q_p = q(np.pi / 2, e)
-    return a * np.sqrt(q_p / 2)
-
-
-def convert(geographic, e):
-    q_ = q(geographic, e)
-    q_p = q(np.pi / 2, e)
-    xi = np.arcsin(q_ / q_p)
-
-    return xi
 
 
 def generate_table():
